fix: update unit test and e2e test
